Working GitLab OAuth authentication
This patch changes the authentication to GitLab OAuth2 and replaces the
login button with a GitLab one. Users will be created on the fly.

The configuration is done via environment variables and an example
is added to the `env.production.sample` file.

You also need to create a new 'application' inside of your GitLab
instance and allow the following callback url:

    https://<your-asciinema-server:port>/auth/gitlab/callback

The created application id and secret have to be added to the
environment file.